requirements.txt
requests
python-dotenv
markdown
pillow
openai
selenium
sentence-transformers

